refactor(messaging): move messaging up to sawtooth
