+++ sawtooth-daml-common/src/main/java/com/blockchaintp/sawtooth/messaging/SendReceiveThread.java
@@ -12,7 +12,7 @@
  limitations under the License.
 ------------------------------------------------------------------------------*/
 
-package com.blockchaintp.sawtooth.daml.messaging;
+package com.blockchaintp.sawtooth.messaging;
 
 import com.google.protobuf.InvalidProtocolBufferException;
