+++ sawtooth-daml-common/src/main/java/com/blockchaintp/sawtooth/messaging/package-info.java
@@ -13,4 +13,4 @@
 /**
  * Zmq Stream implementations.
  */
-package com.blockchaintp.sawtooth.daml.messaging;
+package com.blockchaintp.sawtooth.messaging;
